requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.6
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.9
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.6.0
fsspec==2025.3.2
geomet==0.2.1.post1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
groq==0.23.0
grpcio==1.72.0rc1
grpcio-status==1.63.0rc1
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-google-genai==1.0.10
langchain-groq==0.1.10
langchain-text-splitters==0.2.4
langchainhub==0.1.21
langserve==0.2.3
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.36.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
objectbox==4.0.0
onnxruntime==1.21.1
openai==1.76.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.16
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pillow==11.2.1
posthog==3.25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
